Change animation for nomenclure
